ros2/ubuntu: remove colcon build --parallel-workers argument
If needed, the number of workers can still be limited using

    docker build [...] --cpuset-cpus 0-1

to limit the build containers to only CPU 0 and 1.

nproc would return 2 in this case and the implicit default of colcon/catkin_tools/catkin_make_isolated
is to build with two CPUs then, even if the build host has more than 2 CPUs.