ros2/ubuntu: add COLCON_IGNORE marker in rtt_ros2_integration/tests folder
Even with `-DBUILD_TESTING=OFF` colcon tracks which packages have been built and the generated
top-level setup files expect some per-package setup files in the respective `share/<package>`
directory. But those files are not installed if testing is disabled and the `ament_package()`
CMake macro is not invoked (e.g.
https://github.com/orocos/rtt_ros2_integration/blob/2fe88749a90f7d33051ea58d5d5a0f18853777ff/tests/rtt_ros2_params_tests/CMakeLists.txt#L4-L11).